# todo-ui/src/app/services/models/task-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface TaskResponse {
  createdDate?: string;
  description?: string;
  dueDate?: string;
  id?: number;
  lastUpdatedDate?: string;
  status?: 'CREATED' | 'COMPLETED' | 'CANCELED';
  title?: string;
}